streamlit-deploy/requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0.74
numpy>=1.24.0
requests>=2.31.0